requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
openpyxl==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
